Folder Creation: Working 	File Creation: Working 	Sign/File writing: Working 	Command: Not Working
I have got everything working, but not the command to access the chest.
I am not for sure if it is a server side, because the server I am
testing on is a Spigot server not a Bukkit, but I would think that would
not be a problem. Also I don't know the Bukkit equicalent of Spigot-1517